+++ packages/asciitorium/src/index.ts
@@ -18,6 +18,7 @@ export * from './components/PerfMonitor';
 export * from './components/Line';
 export * from './components/Select';
 export * from './components/Option';
+export * from './components/OptionGroup';
 export * from './components/MultiSelect';
 export * from './components/Tab';
 export * from './components/TabContainer';